Fix bug: file annotations were only shown for current directory
The file annotations where only being displayed for files in the
current directory. This was because the candidates in file completion
are not the full path, but rather just the last path component.
The function file-attributes wants either a full path or a path
relative to the current directory.

To fix this bug I added a general function to calculate the "full
candidate", i.e, what the minibuffer contents would be if a given
candidate is chosen. For most types of completion, the completion
candidates are already full candidates. The main exception is file
name completion, where the candidates are one path component of the
full candidate, which is a full path. (There are other exceptions,
such as environment variable name completion inside file name
completion.)