Go back and use full minibuffer-contents for marginalia--full-file
(Partial revert of 06ae3b4c466801962b7611c851e62bf0c749fbbf)

Otherwise we get oddly disappearing annotations in Vertico when moving the
cursor as observed by @protesilaos in https://github.com/minad/vertico/pull/7.
The argument is that the minibuffer contents are otherwise unchanged and as such
should not affect the computation of the full candidate.

@oantolin, do you strongly disagree with this change?