Fix retrieving full paths under Selectrum
This fix uses Selectrum's selectrum--get-full function. Ideally we
would not depend on this, but it seems tricky to do otherwise.
Selectrum has its own `selectrum-read-file-name` and its own handling
of partial-completion style paths (~/u/s/d/w => ~/usr/sahre/dict/words).

I think we may not be able to find a definition of
marginalia--full-candidate that works under both default completion
and Selectrum, but would be very happy to be proven wrong about that.