Add annotator caching
* Helps when scrolling in incremental UIs
* Does not help much when filtering
* Does not help in non-incremental UIs

Since Marginalia is a major performance offender, I decided to add the caching
here for now. Alternatively the caching layer can be pushed to the affected
completion UIs, in particular Vertico. However the `affixation-function` is a
problem then, since then one has to check the list of candidates first against
the list of already cached candidates and then call the `affixation-function`
once on the new candidates, in order to make only a single call to the
`affixation-function`. Hopefully we can replace the `affixation-function` by an
improved `decoration-function` soon.