annotate variable value when latter is a symbol
Existing behaviour of the variable annotator repeats a variable's name in place of the value when the value is a symbol.  I think this was an unintended part of #90.

This change instead displays the appropriately quoted value of the variable in such cases, and falls back to default print logic in case of a lambda expression value.